pyproject.toml
[tool.poetry]
name = "ai-friedman"
version = "0.1.0"
description = ""
authors = ["Michael Hindley <michael@hindley.se>"]
readme = "README.md"
packages = [{include = "ai_friedman"}]

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.11.0"
langchain = "^0.0.150"
openai = "^0.27.4"
tqdm = "^4.65.0"
pinecone-client = {extras = ["grpc"], version = "^2.2.1"}
tiktoken = "^0.3.3"
pdftotext = "^2.2.2"
nltk = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
